fix(guide): share cover URLs and reuse entities for HTML escape
Move system covers next to the other guide constants and drop the local escapeHtml helper so seed HTML uses the entities package already in sanitize.
