--- app/src/libs/guide/covers.ts
@@ -1,24 +1,0 @@
-/** Hosted cover URLs for first-party system guides. */
-export const GUIDE_SYSTEM_COVERS: Record<string, string> = {
-  "getting-started": "https://ui0arpl8sm.ufs.sh/f/guide-getting-started-v3.webp",
-  combat: "https://ui0arpl8sm.ufs.sh/f/guide-combat-v3.webp",
-  "combat-tags": "https://ui0arpl8sm.ufs.sh/f/guide-combat-tags-v3.webp",
-  "prevent-tags": "https://ui0arpl8sm.ufs.sh/f/guide-prevent-tags-v3.webp",
-  "cleansable-tags": "https://ui0arpl8sm.ufs.sh/f/guide-cleansable-tags-v3.webp",
-  "clearable-tags": "https://ui0arpl8sm.ufs.sh/f/guide-clearable-tags-v3.webp",
-  "combat-tag-priority":
-    "https://ui0arpl8sm.ufs.sh/f/guide-combat-tag-priority-v3.webp",
-  "loadout-building": "https://ui0arpl8sm.ufs.sh/f/guide-loadout-building-v3.webp",
-  "ai-rules": "https://ui0arpl8sm.ufs.sh/f/guide-ai-rules-v3.webp",
-  raids: "https://ui0arpl8sm.ufs.sh/f/guide-raids-v3.webp",
-  "bracket-system": "https://ui0arpl8sm.ufs.sh/f/guide-bracket-system-v3.webp",
-  farming: "https://ui0arpl8sm.ufs.sh/f/guide-farming-v3.webp",
-  villages: "https://ui0arpl8sm.ufs.sh/f/guide-villages-v3.webp",
-  world: "https://ui0arpl8sm.ufs.sh/f/guide-world-v3.webp",
-  "wake-island": "https://ui0arpl8sm.ufs.sh/f/guide-wake-island-v3.webp",
-  bloodlines: "https://ui0arpl8sm.ufs.sh/f/guide-bloodlines-v3.webp",
-  ranks: "https://ui0arpl8sm.ufs.sh/f/guide-ranks-v3.webp",
-  economy: "https://ui0arpl8sm.ufs.sh/f/guide-economy-v3.webp",
-  "auction-house": "https://ui0arpl8sm.ufs.sh/f/guide-auction-house-v3.webp",
-  "item-variants": "https://ui0arpl8sm.ufs.sh/f/guide-item-variants-v3.webp",
-};
